app/requirements.txt
Flask
scipy
matplotlib
numpy
Flask-Cors
Shapely
psycopg2
geopandas
Flask-SQLAlchemy
Flask-Migrate
Flask-Script
py3-ortools
GeoAlchemy2
requests
mapbox
python-dotenv
sklearn
mvg_api